feat: Add start vs join paths and new homepage design
