+++ lib/game_box_web/router.ex
@@ -39,7 +39,8 @@ defmodule GameBoxWeb.Router do
 
     live_session(:default, on_mount: GameBoxWeb.InitAssigns) do
       live("/", HomeLive)
-      live("/join", HomeLive)
+      live("/join", JoinLive)
+      live("/start", StartLive)
       live("/arena/:arena_id", ArenaLive)
       live("/arena/:arena_id/game/:game_id", GameLive)
       live("/signin", SignInLive)
